Revert "Add support for LTTNG_UST_HOME"
This reverts commit 90d125c709f566f3663bf84677f100134cc618e0.

After discussion with Jeremie, we want to introduce two (not one)
environment variables:

- LTTNG_UST_APP_PATH,
- LTTNG_UST_CTL_PATH.

to accomodate use-cases where a sessiond within a container is traced by
a sessiond in the parent container. In that situation, we want the
sessiond in the parent container to access the tracee through the
LTTNG_UST_CTL_PATH, without making the unix sockets for tracing control
visible to the child container.

Therefore, remove the LTTNG_UST_HOME environment variable before it is
added into an official release.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>