Add new flush: flush events or populate packet
Following this change[1], a report was filed[2] on the lttng-dev list
that subsequent snapshots of buffers with no new events would no
longer have events after a number of snapshots equal to the number of
sub-buffers.

The changes were initially made in order to make snapshots on channels
with few changing events intersectable by Babeltrace. By forcibly
closing the packet, the end timestamp represents when the snapshot
was taken and gives the information to the reader that no events were
produced between the last event and the end timestamp of the closed
packet.

Without the above information, the intersection between a busy stream
and a stream with no recent events would make the intersection window
too small.

However, this broke another use case where the content of subsequent
snapshots wouldn't change unless there were new events.

Proposed Solution
-----------------

A new flush function has been added to the public API which will
perform an active flush, but populate a terminal packet that should
also be written to disk or to the network by the consumers in cases
where no change since the last event written have been detected.

This flush is meant to only be used by lttng-tools.

In order to reach this conclusion, the following test is performed
within the new flush operation:

1. The produced position of the stream is sampled
2. The time is sampled
3. An active flush is done
4. The produced position of the stream is sampled again
5. If there there is no change in the produced position of the stream,
it can be inferred that no events have been written into the
sub-buffer. In this case, a packet is populated with the time sampled
in step 2 to give an indication to snapshot readers that no change has
been made between the last event and the timestamp in the popluated
packet.

[1]: https://github.com/lttng/lttng-tools/commit/de3fb857f034c208c135a10a3cdec2dfe43fbda6
[2]: https://lists.lttng.org/pipermail/lttng-dev/2016-December/026805.html

Change-Id: I5a9324354d8bb10e1dafcc3d455508924adf7481
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>