Tests: Fix abi conflict test when building with clang
When building lttng-ust with clang, the following tests in
`tests/regression/abi0-conflict/test_abi0_conflict` fail:

```
not ok 12 - dlopen: no-ust app with abi1 and abi0 fails
not ok 14 - dlopen: ust app with abi0 fails
not ok 15 - dlopen: ust app with abi0 and abi1 fails
not ok 16 - dlopen: ust app with abi1 and abi0 fails
```

Cause
=====

With the default optimizations that are applied, clang seems to inline
the `init_uster` function.

Compiling with `-O0` lets the test pass.

Solution
========

Add the noinline attribute to the `init_uster` function in libfakeust0.

Change-Id: I4a3b16da2c8aad1840cb5e878231b63ca4d47618
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>