doc/examples: set minimal CMake version to 3.5.0
Starting from CMake 4.0.0, the minimal supported version for
`cmake_minimum_required(VERSION ...)` is 3.5.0.

Without this patch and with CMake 4.0.0, you get:

    CMake Error at CMakeLists.txt:5 (cmake_minimum_required):
      Compatibility with CMake < 3.5 has been removed from CMake.

      Update the VERSION argument <min> value.  Or, use the <min>...<max> syntax
      to tell CMake that the project requires at least <min> but has been updated
      to work with policies introduced by <max> or earlier.

      Or, add -DCMAKE_POLICY_VERSION_MINIMUM=3.5 to try configuring anyway.

and the whole build fails.

The workaround is defining `-DCMAKE_POLICY_VERSION_MINIMUM=3.5` [1] when
running CMake, and since this CMake variable is initialized from the
`CMAKE_POLICY_VERSION_MINIMUM` environment variable [2], you can do
something like

    $ CMAKE_POLICY_VERSION_MINIMUM=3.5 make

`cmake_minimum_required(VERSION 3.5.0)` will only work with
CMake ≥ 3.5.0. CMake 3.5.0 was released in December 2018. As of this
date, Ubuntu 20.04 offers CMake 3.16.

Since we didn't have any version check to begin with in `configure.ac`:

    AC_CHECK_PROGS([CMAKE], [cmake])

I'm leaving this as is as this would be a different concern.

[1]: https://cmake.org/cmake/help/latest/variable/CMAKE_POLICY_VERSION_MINIMUM.html
[2]: https://cmake.org/cmake/help/latest/envvar/CMAKE_POLICY_VERSION_MINIMUM.html

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: Iacfa07da5775eca979aeb34ec5d320d83a82a5dc