ABI: Bump major version to 11
The ABI needs to be bump because transmitted channels now require a
`owner_id' field to be present.

This new version also introduces the concept of sub-buffer ownership,
where producers are expected to take the ownership of sub-buffers before
making a reservation. Furthermore, the lazy-padding concept is
introduced. This guarantees that only a single hole can exist in a
sub-buffer and is at the end.

Change-Id: I2c86ce39692b2a1fe8020f4db8dfb4218fc9b7b7
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>