ring-buffer: Add discarded events at packet header beginning
The stalled buffers fixup algorithm sometime requires that a next
sub-buffer exists. For example, when the packet header end timestamp is
missing, it is copied from the packet header begin timestamp of the next
sub-buffer.

The same logic applies for discarded events counter. But, this counter
is only emitted when ending the packet (at the same moment as the end
timestamp).

Therefore introduce the discarded events counter emitted when beginning
a packet. However, because this change requires changing viewers like
Babeltrace, the counter is not actually written in the packet header
yet, but is instead placed in the cold commit counter structure of the
sub-buffer, thus keeping the value internal to UST for now.

Change-Id: I82199deed19e5d67baa0468417465021e721be0a
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>