ring-buffer: Implement lazy-padding
The lazy padding algorithm works as follow:

Before making a reservation, a producer needs to take the ownership of
the targeted sub-buffer by successfully exchanging its owner-id, while
expecting the unset value to be present. Then, the reservation can be
made followed by the commit. Before releasing the ownership of the
sub-buffer, the producer checks if the reserve position has moved past
its owned sub-buffer. In that case, the producer needs to add a padding
hole to the sub-buffer and then deliver it.

The lazy padding algorithm is required for robust shared memory
transactions. It allows an external fixer (typically the consumer) to
fix an unfinished memory transaction due to a crash of a producer. The
fix itself is not introduced in this commit, instead it setup the state
machine required for being able to fix transactions.

The impact of the algorithm on the fast path is an additional compare
exchange operation on the `owner_id' field of the sub-buffer before
taking the reservation. It also adds an additional store release for
releasing the ownership of the sub-buffer after committing.

Furthermore, the hot and cold commit counters are now protected by the
ownership of the sub-buffer and can thus be mutated with volatile access
by the owner of the sub-buffer.

Change-Id: Ibc76acc553bfc4c05f0b7e188f508d31f120ac4e
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>