ring-buffer: Use lazy-padding
This commit breaks the ring buffer shared memory ABI, and is therefore
introduced with a bump of the ABI major number.

Ownership is now taken for the sub-buffer and the lazy-padding is done
when necessary.

This commit also introduces testpoints with detailed explanation of the
observable side-effects associated with them. They are used by
regression tests to position producers at very precise locations in the
state machine and crashing them. The goal is to have as many as possible
reproducible memory transaction failure scenarios while not brute
forcing race conditions.

Change-Id: Icf00e7252f3dbdd84ada92f04d92825e38f4da5a
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>