ring-buffer: Speed up pushing reader consumed position ahead
Instead of pushing the reader a single sub-buffer size at a time, push
the reader to the maximum distance it can be from the produced position
directly.

Historically, the consumed position was just moved one sub-buffer
ahead. But because of the new stalled fixup algorithm, it is faster to
just move the consumed position to the next valid position directly.

Change-Id: Iff56d4c06eaf7bf0e4584ceaae4d5a96b8e2b758
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>