ring-buffer: Implement packet count as sequence-counters
The packet counts is a 64-bit counter that is incremented during a
deliver, under the protection of the increment of the cold commit
counter, which acts as a mutual exclusion on the sub-buffer.

However, this operation needs to be restartable by a stalled fixup and
observable under the full balance of the cold commit counter and the hot
commit counter.

This can be achieved by making the packet count counter a
sequence-counter with two slots, where the current slot index is derived
from the cold commit counter.

Change-Id: I9fc747a14ea8aa40c40a3cb154c9d1db097a5fc2
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>