ust-ctl: Implement stalled buffers fixup
This commit exposes the fixup algorithm. It allows fixing memory
transactions that are stalled due to dead owners, i.e. owners that
crashed during a memory transaction.

This has the effect of being able to destroy a tracing session and get
the full trace even when producers crashed while emitting events.

Before this, if a producer was taking a reservation in the ring-buffer
and crashed before committing, the ring-buffer was in a unbalanced state
that results in:

  1. All events being dropped when wrapping around to the faulty
  sub-buffer

  2. Destroying a session was waiting endlessly for the ring-buffer to
  be balanced.

Change-Id: I8be8b58397cfa0d2ae483f272284952d2516de11
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>