Fix: Move begin_events_discarded to backend_pages
In high-throughput tests, in flight recorder tracing, the content
of the commit_cold structure can be overwritten by the next producer
pass even though the consumer has done a successful "get_subbuf".

This is because the commit_cold structure is associated with the
per-writer subbuffer index (wsb_index) used by the ring buffer frontend.

Our goal here is to read the stored begin_events_discarded value
from a consumerd reader. We therefore need to move this field from
commit_cold to the backend_pages, which are implicitly moved around
with the subbuffer memory when exchanging the subbuffer indexes.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I726f65ffda4a551724b4f057e16624546448ffb1
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>