ring-buffer: support optional lazy allocation of backing storage
Introduce a `preallocate_backing` parameter to control whether
ring-buffer backing storage is allocated up front or on demand.

When `preallocate_backing` is set to false, ring buffers are created
using sparse files and backing storage is allocated lazily as data is
written.

This is useful when some CPUs may never write to their ring buffer.
Some examples include:
- When the ring-buffer is used by a containerized application that may
not use all available CPUs,
- When a per-PID ring-buffer is allocated for an application that is
affined to a limited number of CPUs.

Note that since follow-up changes introduce other uses of fallocate()
(only available on Linux), the strategy for allocating backing
storage now benefits from using fallocate() if available and supported
by the underlying filesystem and falls back to zeroing the file if it is
not.

Also, for this specific use of fallocate(), it would be possible to
use the more portable posix_fallocate() instead of the Linux-specific
fallocate().

Moreover, since the lttng_ust_ctl_consumer_channel_attr structure's
layout is changed, the "packed" attribute is removed as it appears to be
unnecessary (the structure no longer appears to be used "on the wire").

** Should the same option be provided for counters? **

After the allocation of the ring-buffers, the size of the backing files
isn't reported as zero. This is expected since the memory layout of the
ring-buffers includes a number of control structures (the "crash" ABI
description, commit counters, backend_pages, etc.)

For reference, here's the allocated size reported by `stat` after the
initial allocation of a single ring-buffer:

| Configuration                   | Reported Size | Size Overhead    |
|---------------------------------|---------------|------------------|
| 4 sub-buffers, 512 KiB each     | 12,288 bytes  |     ~0.59%       |
| 4096 sub-buffers, 4 KiB each    | 376,832 bytes |     ~2.2%        |

Change-Id: I01f7a97479d13fc7033c50a8ad0b8d296b225a6d
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>