subbuf_iter: Report the sub-buffer's allocated state instead of populated
A buffer has two states that are subtly different: allocated and
populated.

The `allocated` state refers to the backing storage's state.
Effectively, it informs the user about whether or not the sub-buffer is
being backed by a "hole" in a sparse file or if it backed by allocated
pages.

The `populated` state refers to whether or not the page table of a
buffer has been prefaulted (see mmap(2), MAP_POPULATE).

For the moment, the users of the subbuf_iter API are concerned with the
`allocated` state of the sub-buffers to determine if a reclamation
should be attempted to reduce the memory usage via the use of a sparse
file.

The public API is modified to reflect that change and the internal
terminology is clarified in a number of places.

Change-Id: Ie0fd280bab4d2ea85b0cd1b6ddc35ffb80739210
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>