Don't produce an initial packet header on ring-buffer creation
The first sub-buffer of a ring-buffer is always initialized with
an initial packet header. This was originally done to provide the
buffer's creation timestamp.

From a reader's perspective, this allows it to know the beginning of the
event recording period: from that point on, if an event of interest had
occurred, it would either be recorded in the buffer, or it would be
recorded as a discarded event, or it would result in a packet being
dropped, depending on the event record loss mode.

Also, the `timestamp_begin` value matching the start of the recording
session is key to the "stream intersection" functionality of Babeltrace.
Without that, it becomes more or less unusable since the intersection
would become extremely short whenever a ring-buffer has seen little
activity during the session's lifetime.

However, doing so causes the first buffer of every ring-buffer to appear
"in-use" preventing them from being reclaimed. It also results in
the allocation of a page for each sub-buffer even when "on-demand"
allocation is specified as part of the channel's configuration.

To reduce the memory usage and preserve the current user-visible
behaviour, the header is only inserted when entering the sub-buffer on
the first go-around based on a new lttng_ust_ring_buffer flag:
`use_creation_timestamp`. The header then uses the channel backend's
`start_timestamp`.

Change-Id: Ibd2a24e0b9e0d9d1efd305e962156c3f20f00540
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>