Move lttng_ust_abi_channel_config to ust-abi.h, cleanup layout
There is a TODO for cleaning up this ABI structure that was awaiting a
major version bump of the communication protocol. Do it now that we
are bumping the version.

Move the structure to ust-abi.h to clarify that this is indeed ABI.

Add padding for future extension.

Remove deprecated and unused fields.

Rename LTTNG_UST_UUID_LEN to LTTNG_UST_ABI_UUID_LEN.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I974dde37fc12144c5a7d54d1985e6a170f32fde1