Refactor the LTTng-UST communication protocol
Use the UST protocol bump to refactor the communication protocol
between the session and consumer daemons and lttng-ust within
instrumented applications.

Replace the current scheme which is based on packed struct with
fixed-length fields and a following padding field by a scheme where a
padding field is placed in a union alongside with an anonymous structure
that contains the payload. This allows making the size of each message
explicit, and validate them with static assertions.

GCC 15+ only zero-initializes the first union member. Move the padding
field to first fields to ensure that the entire structure is
initialized.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: If68ddfb45772531dcf9a33620256306de373abea