testpoint: Ensure unique testpoint labels
Some testpoints are emitted in inline functions.  This is a problem
because such testpoint ASM statement is expanded multiple times in the
object file.

For example, consider this reproducer:

  #define TESTPOINT(label) \
    asm (".local " label "\n\t" label "= .\n\t" : : : )

  static inline void foo(int *x, int y)
  {
      __atomic_store_n(x, y, __ATOMIC_SEQ_CST);
      TESTPOINT("TESTPOINT");
  }

  void f(int *x, int y)
  {
      foo(x, y);
  }

  void f2(int *x, int y)
  {
      foo(x, y);
      foo(x, y);
  }

and its resulting object dump:

  f:
   xchg   %esi,(%rdi)
   ret
   data16 cs nopw 0x0(%rax,%rax,1)
   xchg   %ax,%ax
  f2:
   mov    %esi,%eax
   xchg   %eax,(%rdi)
   xchg   %esi,(%rdi)
  TESTPOINT:
   ret

The TESTPOINT label is only emitted once when we would expect it to be
emitted 3 times.  This could be solved by adding a unique number suffix
to the testpoints.

The `%=' special format can just do that in _extended_ ASM template
only.  Furthermore, to avoid the compiler re-ordering the testpoints
across potentially other ASM statements, the volatile qualifier is used.

Change-Id: I50f532afff02d62413dc65d2d92d7ea60776f4a2
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>