Introduce lttng_ust_ctl_timestamp_sync
Issue
==============

Using the stream creation time to lazily stamp the first packet within a
stream is problematic for live sessions. An inactive stream will have
inactivity beacons claiming no activity up to a given timestamp, but can
then emit the timestamp_begin with the creation timestamp, which is
prior to the last inactivity beacon timestamp. This is inconsistent
from the perspective of the lttng live protocol.

Solution
========

Introduce a lttng_ust_ctl_timestamp_sync() API, which allows the live
beacon to synchronize time for this stream by ensuring it will not use
the creation time when it emits the first packet for the stream in the
future.

This is done by emitting an empty packet (SWITCH_FLUSH) if the
use_creation_timestamp flag is still set for the buffer.

Known drawbacks
===============

Similarly to "lttng stop", the first live beacon timer will require
allocation of one backing page per CPU in the system for each inactive
stream. This burst of memory use is temporary if the memory is reclaimed
by the consumer daemon afterward.

Fixes: 81b931f1da86 ("Don't produce an initial packet header on ring-buffer creation")
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: Ide3b82c88ac527946400c9e1e89662700d837a00