lttng_ust_ctl_timestamp_add: clamp timestamp value to 0
lttng_ust_ctl_timestamp_add can be used with a negative delta to
decrement a timestamp value. If this negative delta is larger than the
original value, this underflows the 64-bit unsigned integer.

Deal with underflow by clamping the resulting timestamp to 0.

This deals with a scenario where a user requests reclaim of sub-buffer
memory older than 24h, and starts tracing a few minutes after boot. The
resulting timestamp to compare with therefore underflows an unsigned
64-bit integer. Clamp the comparison value to 0 instead.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I994a8b6f58b64ecf1b79b9c95f6fe2c552b5019d