Fix: incorrect subbuffer ID initialization
A subbuffer ID is made of three parts: an "offset", a no_ref flag, and
an "index".

On 32-bit architectures:
  - offset (15 upper bits) | no_ref (1 bit) | index (16 lower bits)

On 64-bit architectures:
  - offset (31 upper bits) | no_ref (1 bit) | index (32 lower bits)

On 32-bit architectures, when attempting to reclaim the memory of a
ring-buffer that saw no producer activity, the
lib_ring_buffer_try_exchange_subbuf would fail systematically on the
subbuffer_id_compare_offset check.

The position provided by the subbuf_iterator has a logically negative
value since the "produced" position is negative. At that point, we are
in "pre-roll". When that value is transformed into a subbuffer ID, the
upper bits of the position are set to 0 when performing an unsigned
right shift (see buf_trunc_val).

Assuming a 2MiB ring-buffer, 21 bits are needed to represent the offset
inside a buffer "iteration". Hence, to get the number of completed
buffer iterations from the offset, it is shifted right by 21.

This unsigned shift clears the 21 upper bits leaving in place 11 enabled
bits.

This is compared to the wsb_id offset which consists of 15 enabled bits
(for 32-bit architectures).

Since the comparison of the offsets happens on 15 bits (see
SB_ID_OFFSET_MASK and subbuffer_id_compare_offset), the comparison
fails.

** Why it works on 64-bit architectures **

Assuming a 2MiB ring-buffer, 21 bits are needed to represent the offset
inside a buffer "iteration". Hence, to get the number of completed
buffer iterations from the offset, it is shifted right by 21.

This unsigned shift clears the 21 upper bits leaving in place *43*
enabled bits.

This is compared to the wsb_id offset which consists of 31 enabled bits.

Since the comparison of the offsets happens on 31 bits (see
SB_ID_OFFSET_MASK and subbuffer_id_compare_offset), it succeeds in
comparing the limited range of values.

The fix consists in restricting the number of bits used for the offset
to those left after right shifting by the amount needed to represent a
position in one iteration of the ring-buffer.

Fixes: 409093bebaa3 ("Fix: initialize sub-buffers's id to a pre-roll loop")
Change-Id: Ieb93390e32fc46ebffd2e56b3d1d25a569749347
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>